// src/app/(root)/ctgom/interface.ts
export interface Usuario {
  id: number;
  nome: string;
  email: string;
  telefone: string;
  bi: string;
  data_nascimento: string; // "YYYY-MM-DD"
  genero: string;
  status: string;
  tipo: string;
  natural_id: number | null;
  criado_por: number | null;
  created_at: string;
  updated_at: string;
  deleted_at: string | null;
}